Editing File Structure
