+++ lib/screens/search_page.dart
@@ -2,11 +2,11 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:notify/components/methods/snapshot_middleware.dart';
 import 'package:notify/components/widgets/notify_text_field.dart';
 import 'package:notify/components/widgets/notify_user_list_tile.dart';
+import 'package:notify/services/classes/notify_user.dart';
 import 'package:notify/services/firebase_service.dart';
-import 'package:notify/services/notify_user.dart';
+import 'package:notify/static_methods/snapshot_middleware.dart';
 import 'package:provider/provider.dart';
 
 class SearchPage extends StatefulWidget {
@@ -70,8 +70,7 @@ class _SearchPageState extends State<SearchPage>
                 );
               }
               return StreamBuilder<List<NotifyUser>>(
-                stream: context
-                    .read<FirebaseService>()
+                stream: FirebaseService.of(context)
                     .getUsersListFromUsersUidList(state),
                 builder: (
                   final BuildContext context,
