+++ lib/static_methods/combine_latest_streams.dart
