+++ lib/static_methods/snapshot_middleware.dart
